+++ sat/cli/status/main.py
@@ -130,5 +130,7 @@ def do_status(args):
 
     if args.format == 'yaml':
         print(report.get_yaml())
+    elif args.format == 'json':
+        print(report.get_json())
     else:
         print(report)