+++ CHANGELOG.md
@@ -27,6 +27,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+- Fixed a case where `sat xname2nid` and `sat nid2xname` failed when the input had
+  extraneous whitespace
+
 ### Added
 - Added a ``--fields`` option to allow displaying only specific fields
   in subcommands which display a report.