+++ CHANGELOG.md
@@ -36,6 +36,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   extraneous whitespace.
 - Fixed `--logfile` to accept a file name without leading directories.
 - Fixed issues with filter error messages being printed twice or not being printed.
+- Improved error messages when invalid filter syntax was supplied to
+  refer to documentation of the syntax specification.
 
 ### Added
 - Added a ``--fields`` option to allow displaying only specific fields